Fixed crash when a subreddit with no images is provided

+++ subreddit.py
@@ -29,6 +29,7 @@ def subredditscript(command, client, message):
             postcount += 1
         print("Post range = ",postcount)
         
+        w=0
         x=True
         accepted_type= ["jpg","peg","png"] #last three characters of image file extensions
         while x:
@@ -43,21 +44,32 @@ def subredditscript(command, client, message):
             if url_type in accepted_type:
                 x=False
                 print("URL is to an image file")
-        if url_type=="jpg":
-            file_name="images/"+str(random.randint(1,1000))+".jpg"
-        if url_type=="jpeg":
-            file_name="images/"+str(random.randint(1,1000))+".jpeg"
-        if url_type=="png":
-            file_name="images/"+str(random.randint(1,1000))+".png"
-        if url_type=="gif":
-            file_name="images/"+str(random.randint(1,1000))+".gif"            
-        print("Downloading image from: ", url)
-        urllib.request.urlretrieve(url, file_name)
-        open(file_name + ".lock","a").close()
-        print("Image placed at: ",file_name," and locked")
-        time.sleep(1)
-        print("Sending Image")
-        run_coro(client.send_file(message.channel, file_name), client) # Send Image
-        print("Sent Image")
-        os.remove(file_name + ".lock")
-        print(file_name, " unlocked.")
+            w += 1
+            if w==5:
+                x=False
+                run_coro(client.send_message(message.channel, "An image could not be found in this subreddit."), client)
+                foundimage=0
+            else:
+                foundimage=1
+                
+                
+        if foundimage==1:
+            print("Image Found")
+            if url_type=="jpg":
+                file_name="images/"+str(random.randint(1,1000))+".jpg"
+            if url_type=="jpeg":
+                file_name="images/"+str(random.randint(1,1000))+".jpeg"
+            if url_type=="png":
+                file_name="images/"+str(random.randint(1,1000))+".png"
+            if url_type=="gif":
+                file_name="images/"+str(random.randint(1,1000))+".gif"            
+            print("Downloading image from: ", url)
+            urllib.request.urlretrieve(url, file_name)
+            open(file_name + ".lock","a").close()
+            print("Image placed at: ",file_name," and locked")
+            time.sleep(1)
+            print("Sending Image")
+            run_coro(client.send_file(message.channel, file_name), client) # Send Image
+            print("Sent Image")
+            os.remove(file_name + ".lock")
+            print(file_name, " unlocked.")
